+++ src/main/java/com/spata14/trelloproject/board/service/BoardService.java
@@ -6,7 +6,7 @@ import com.spata14.trelloproject.board.dto.BoardResponseDto;
 import com.spata14.trelloproject.board.entity.Board;
 import com.spata14.trelloproject.board.repository.BoardRepository;
 import com.spata14.trelloproject.user.User;
-import com.spata14.trelloproject.user.UserRole;
+import com.spata14.trelloproject.workspace.WorkspaceMemberRole;
 import com.spata14.trelloproject.user.repository.UserRepository;
 import com.spata14.trelloproject.workspace.Workspace;
 import com.spata14.trelloproject.workspace.WorkspaceUser;
@@ -44,7 +44,7 @@ public class BoardService {
                 .orElseThrow(() -> new RuntimeException("해당 유저는 워크스페이스의 멤버가 아닙니다"));
 
         // 권한 검사: READ 권한은 보드 생성 불가
-        if (workspaceUser.getUserRole() == UserRole.READ) {
+        if (workspaceUser.getWorkspaceMemberRole() == WorkspaceMemberRole.READ_ONLY) {
             throw new RuntimeException("읽기 권한을 가진 유저는 보드 생성이 불가능합니다");
         }
 
@@ -91,7 +91,7 @@ public class BoardService {
                 .orElseThrow(() -> new RuntimeException("해당 유저는 워크스페이스의 멤버가 아닙니다"));
 
         // 권한 검사: READ 권한은 보드 수정 불가
-        if (workspaceUser.getUserRole() == UserRole.READ) {
+        if (workspaceUser.getWorkspaceMemberRole() == WorkspaceMemberRole.READ_ONLY) {
             throw new RuntimeException("읽기 권한을 가진 유저는 보드를 수정할 수 없습니다.");
         }
