--- src/main/java/com/spata14/trelloproject/user/UserRole.java
@@ -1,16 +1,0 @@
-package com.spata14.trelloproject.user;
-
-import lombok.Getter;
-
-@Getter
-public enum UserRole {
-    NONE("none"),
-    READ("read"),
-    ALL("all");
-
-    private final String name;
-
-    UserRole(String name) {
-        this.name = name;
-    }
-}
