+++ src/main/java/com/spata14/trelloproject/workspace/WorkspaceMemberRole.java
@@ -1,0 +1,16 @@
+package com.spata14.trelloproject.workspace;
+
+import lombok.Getter;
+
+@Getter
+public enum WorkspaceMemberRole {
+    READ_ONLY("readOnly"),
+    BOARD("board"),
+    WORKSPACE("workspace");
+
+    private final String name;
+
+    WorkspaceMemberRole(String name) {
+        this.name = name;
+    }
+}
