+++ src/main/java/com/spata14/trelloproject/workspace/dto/InviteMemberRequestDto.java
@@ -1,5 +1,6 @@
 package com.spata14.trelloproject.workspace.dto;
 
+import com.spata14.trelloproject.workspace.WorkspaceMemberRole;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -7,4 +8,5 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class InviteMemberRequestDto {
     private final String email;
+    private final WorkspaceMemberRole workspaceMemberRole;
 }
