+++ src/main/java/com/spata14/trelloproject/workspace/repository/WorkspaceUserRepository.java
@@ -1,6 +1,6 @@
 package com.spata14.trelloproject.workspace.repository;
 
-import com.spata14.trelloproject.user.UserRole;
+import com.spata14.trelloproject.workspace.WorkspaceMemberRole;
 import com.spata14.trelloproject.workspace.WorkspaceUser;
 import com.spata14.trelloproject.workspace.exception.WorkspaceErrorCode;
 import com.spata14.trelloproject.workspace.exception.WorkspaceException;
@@ -16,18 +16,18 @@ public interface WorkspaceUserRepository extends JpaRepository<WorkspaceUser, Lo
             " join fetch wu.user" +
             " join fetch wu.workspace" +
             " where wu.user.email = :email" +
-            " and wu.userRole = :userRole" +
+            " and wu.workspaceMemberRole = :workspaceMemberRole" +
             " and wu.workspace.id = :workspaceId")
-    Optional<WorkspaceUser> getWorkspaceOwner(@Param("email") String email, @Param("userRole")UserRole userRole, @Param("workspaceId") Long workspaceId);
+    Optional<WorkspaceUser> getWorkspaceOwner(@Param("email") String email, @Param("workspaceMemberRole") WorkspaceMemberRole workspaceMemberRole, @Param("workspaceId") Long workspaceId);
 
     /**
      *
      * @param email - 유저 이메일
-     * @param userRole - 워크스페이스 내 유저 역할
+     * @param workspaceMemberRole - 워크스페이스 내 유저 역할
      * @param workspaceId - 워크스페이스 ID
      * @return {@link WorkspaceUser}
      */
-    default WorkspaceUser getWorkspaceOwnerOrElseThrow(String email, UserRole userRole, Long workspaceId) {
-        return getWorkspaceOwner(email, userRole, workspaceId).orElseThrow(() -> new WorkspaceException(WorkspaceErrorCode.WORKSPACE_UNAUTHORIZED));
+    default WorkspaceUser getWorkspaceOwnerOrElseThrow(String email, WorkspaceMemberRole workspaceMemberRole, Long workspaceId) {
+        return getWorkspaceOwner(email, workspaceMemberRole, workspaceId).orElseThrow(() -> new WorkspaceException(WorkspaceErrorCode.WORKSPACE_UNAUTHORIZED));
     }
 }
