+++ src/main/java/com/spata14/trelloproject/workspace/service/WorkspaceService.java
@@ -2,7 +2,7 @@ package com.spata14.trelloproject.workspace.service;
 
 import com.spata14.trelloproject.common.util.SessionUtils;
 import com.spata14.trelloproject.user.User;
-import com.spata14.trelloproject.user.UserRole;
+import com.spata14.trelloproject.workspace.WorkspaceMemberRole;
 import com.spata14.trelloproject.user.service.UserService;
 import com.spata14.trelloproject.workspace.Workspace;
 import com.spata14.trelloproject.workspace.WorkspaceUser;
@@ -41,8 +41,8 @@ public class WorkspaceService {
         Workspace workspace = new Workspace(dto.getName(), dto.getDescription());
         workspaceRepository.save(workspace);
 
-        //생성자 역할을 ALL 로 설정 (admin)
-        workspaceUserRepository.save(new WorkspaceUser(mySelf, workspace, UserRole.ALL));
+        //생성자 역할을 WORKSPACE 로 설정
+        workspaceUserRepository.save(new WorkspaceUser(mySelf, workspace, WorkspaceMemberRole.WORKSPACE));
 
         return WorkspaceResponseDto.toDto(workspace);
     }
@@ -50,13 +50,13 @@ public class WorkspaceService {
     /**
      * 멤버 추가
      * @param id 워크스페이스 ID
-     * @param dto {@link InviteMemberRequestDto} - email(String)
+     * @param dto {@link InviteMemberRequestDto} - email(String), workspaceMemberRole(Enum) WORKSPACE, BOARD, READ_ONLY
      * @return String (user.getEmail() + " 님을 워크스페이스에 초대하였습니다.")
      */
     @Transactional
     public String addMember(Long id, InviteMemberRequestDto dto) {
-        // 워크스페이스 소유자 체크
-        WorkspaceUser workspaceUser = workspaceUserRepository.getWorkspaceOwnerOrElseThrow(sessionUtils.getLoginUserEmail(), UserRole.ALL, id);
+        // 워크스페이스 역할 체크
+        WorkspaceUser workspaceUser = workspaceUserRepository.getWorkspaceOwnerOrElseThrow(sessionUtils.getLoginUserEmail(), WorkspaceMemberRole.WORKSPACE, id);
 
         // 초대 중복 방지
         if (workspaceUserRepository.isMember(id, dto.getEmail())) {
@@ -66,8 +66,8 @@ public class WorkspaceService {
         // 멤버로 추가할 유저
         User user = userService.findUserByEmail(dto.getEmail());
 
-        // 역할을 READ 로 설정
-        workspaceUserRepository.save(new WorkspaceUser(user, workspaceUser.getWorkspace(), UserRole.READ));
+        // 워크 스페이스에 유저 저장, 역할 지정
+        workspaceUserRepository.save(new WorkspaceUser(user, workspaceUser.getWorkspace(), dto.getWorkspaceMemberRole()));
 
         return user.getEmail() + " 님을 워크스페이스에 초대하였습니다.";
     }
@@ -88,7 +88,7 @@ public class WorkspaceService {
      */
     @Transactional
     public WorkspaceResponseDto update(Long id, WorkspaceRequestDto dto) {
-        WorkspaceUser workspaceUser = workspaceUserRepository.getWorkspaceOwnerOrElseThrow(sessionUtils.getLoginUserEmail(), UserRole.ALL, id);
+        WorkspaceUser workspaceUser = workspaceUserRepository.getWorkspaceOwnerOrElseThrow(sessionUtils.getLoginUserEmail(), WorkspaceMemberRole.WORKSPACE, id);
         Workspace workspace = workspaceUser.getWorkspace();
         workspace.updateWorkspace(dto.getName(), dto.getDescription());
         workspaceRepository.save(workspace);
@@ -102,7 +102,7 @@ public class WorkspaceService {
      */
     @Transactional
     public void delete(Long id) {
-        WorkspaceUser workspaceUser = workspaceUserRepository.getWorkspaceOwnerOrElseThrow(sessionUtils.getLoginUserEmail(), UserRole.ALL, id);
+        WorkspaceUser workspaceUser = workspaceUserRepository.getWorkspaceOwnerOrElseThrow(sessionUtils.getLoginUserEmail(), WorkspaceMemberRole.WORKSPACE, id);
         workspaceRepository.delete(workspaceUser.getWorkspace());
     }
 }
